+++ src/Console/Application.php
@@ -118,6 +118,7 @@ EOF;
         $commands[] = new \SugarCli\Console\Command\Anonymize\AnonymizeRunCommand();
         $commands[] = new \SugarCli\Console\Command\Backup\DumpFilesCommand();
         $commands[] = new \SugarCli\Console\Command\Backup\DumpDatabaseCommand();
+        $commands[] = new \SugarCli\Console\Command\Backup\DumpAllCommand();
         $commands[] = new \SugarCli\Console\Command\Backup\RestoreFilesCommand();
         $commands[] = new \SugarCli\Console\Command\CleanLangFilesCommand();
         $commands[] = new \SugarCli\Console\Command\Code\ButtonCommand();