+++ CHANGELOG.md
@@ -3,9 +3,10 @@ Changelog
 
 (unreleased)
 ----
+* Add command `backup:dump:all`
+* Add command `backup:dump:database`
 * Add command `backup:dump:files`
 * Add command `backup:restore:files`
-* Add command `backup:dump:database`
 
 1.15.1
 ----